+++ lib/engineyard-recipes/generators/sm_generator.rb
@@ -9,7 +9,6 @@ module Engineyard::Recipes
       argument :recipe_name
       argument :sm_ext_uri
       argument :sm_ext_commands, :type => :array
-      argument :flags, :required => false
 
       def self.source_root
         File.join(File.dirname(__FILE__), "sm_generator", "templates")
@@ -35,21 +34,15 @@ module Engineyard::Recipes
       end
       
       def vendor
-        if submodule?
-          FetchUri.vendor_recipe_into_recipe(sm_ext_uri, sm_vendor_path)
-        end
+        FetchUri.vendor_recipe_into_recipe(sm_ext_uri, recipe_vendor_path)
       end
       
       protected
-      def submodule?
-        flags[:submodule]
-      end
-      
-      def sm_vendor_path_name
+      def recipe_vendor_path_name
         "repo"
       end
       
-      def sm_vendor_path
+      def recipe_vendor_path
         cookbooks_dir File.join(recipe_name, "repo")
       end
     end